Time: 4 ms (100.00%), Space: 43.8 MB (63.32%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.java
@@ -63,7 +63,7 @@ class Solution {
         
         
         int dig = 0;
-        int reverse = 0;
+        long reverse = 0;
         int tempx = x;
         
         
@@ -77,7 +77,7 @@ class Solution {
             
         }
         
-        if(reverse==x){
+        if((int)reverse==x){
             return true;
         }
         return false;
